Initial version of launcher jar. Installer now unpacks the launcher jar and then starts it.
